bot/requirements.txt
aiogram==3.3.0
aiohttp==3.9.1
openai==1.3.7
sentry-sdk==1.38.0
python-dotenv==1.0.0
aiofiles==23.2.1
httpx==0.25.2
